clock-beat-client/requirements.txt
pygame~=2.6.1
pyserial~=3.5
requests~=2.32.4